:bug: Fix `prop_add_list` function on Vim
