+++ denops_std/function/vim/_manual.ts
@@ -1,0 +1,1 @@
+export * from "./prop_add_list.ts";
